+++ lib/generators/administrate/manifest/templates/dashboard_manifest.rb.erb
